Pushed Tor interface to external class

+++ tor-scraper.py
@@ -3,64 +3,24 @@
 # NOTE: This version requires Python 3.x to run
 
 import sys, os
-import urllib
 import socket
-import http
 import re
 import argparse
 import datetime
 import queue
 import concurrent.futures
 
-import socks		# Now using branch version for compatibility with Python 3.x
-			# https://code.google.com/p/socksipy-branch/
-
-import stem		# http://stem.torproject.org
-import stem.process
-import stem.control
-from stem.util import term
-
 import couchdb		# Running branch patched for Python 3
 			# https://github.com/lilydjwg/couchdb-python3
 from couchdb.mapping import Document, TextField, IntegerField, DateTimeField, BooleanField
+
+import TorHandler
 
 # Enable or disable debug messages
 # Codes:
 # [I] - Informational Messages
 # [E] - Error Messages
 DEBUG = True
-
-def query(url):
-	try:
-		opener = urllib.request.build_opener(SocksiPyHandler(socks.PROXY_TYPE_SOCKS5, "localhost", 9050))
-		h = opener.open(url)
-		return h
-	except:
-		print("[E] Unable to return {0}".format(url))
-
-class SocksiPyConnection(http.client.HTTPConnection):
-    def __init__(self, proxytype, proxyaddr, proxyport = None, rdns = True, username = None, password = None, *args, **kwargs):
-        self.proxyargs = (proxytype, proxyaddr, proxyport, rdns, username, password)
-        http.client.HTTPConnection.__init__(self, *args, **kwargs)
-
-    def connect(self):
-        self.sock = socks.socksocket()
-        self.sock.setproxy(*self.proxyargs)
-        if isinstance(self.timeout, float):
-            self.sock.settimeout(self.timeout)
-        self.sock.connect((self.host, self.port))
-
-class SocksiPyHandler(urllib.request.HTTPHandler):
-    def __init__(self, *args, **kwargs):
-        self.args = args
-        self.kw = kwargs
-        urllib.request.HTTPHandler.__init__(self)
-
-    def http_open(self, req):
-        def build(host, port=None, strict=None, timeout=0):
-            conn = SocksiPyConnection(*self.args, host=host, port=port, strict=strict, timeout=timeout, **self.kw)
-            return conn
-        return self.do_open(build, req)
 
 class DB_Structure(Document):
 	_id = TextField()
@@ -70,11 +30,6 @@ class DB_Structure(Document):
 	LastAccessed = DateTimeField()
 	title = TextField()
 	is_alive = BooleanField()
-
-
-def print_bootstrap_lines(line):
-	if "Bootstrapped" in line:
-		print(term.format(line, term.Color.BLUE))
 
 def check_http(url):
 	# urllib needs URLs to include http://, so if the user submits "example.com" this will prepend http://
@@ -91,30 +46,7 @@ def get_title(lines):
 			title = m.group(0) 
 	return title[7:-8]
 
-def start_tor():
-	print(term.format("Starting Tor:\n", term.Attr.BOLD))
-	try:
-		tor_process = stem.process.launch_tor_with_config(
-			config = {
-				'SocksPort':str(9050),
-				'ControlPort':str(9051),
-				},
-			init_msg_handler = print_bootstrap_lines,
-			completion_percent = 100,
-			timeout = 3600,
-			take_ownership = True
-		)
-
-	# TODO: implement the ability to bind to existing Tor instance in the future
-	# use stem.connection lib
-	except OSError as e:
-		print("[E] There was an error launching Tor. It may already be running.")
-		if DEBUG:
-			print("Error message: {0}".format(e))
-		# TODO: Remove this, make script try to connect to running instance
-		#sys.exit("Please kill all running Tor instances and try again")
-
-def scrape_site(site, domains, db):
+def scrape_site(site, domains, db, handler):
 	doc = DB_Structure.load(db, check_http(site))
 	if doc.LastAccessed is not None and (datetime.datetime.now() - doc.LastAccessed).total_seconds() < 300:
 		if DEBUG:
@@ -123,10 +55,10 @@ def scrape_site(site, domains, db):
 
 	alive = True
 	if DEBUG:
-		print(term.format("Scraping %s\n" % site, term.Attr.BOLD))
+		print("Scraping {0}".format(site))
 
 	try:
-		lines = query(site).readlines()
+		lines = handler.query(site).readlines()
 
 	except AttributeError as e:
 		print("[E] Site could not be read!")
@@ -173,19 +105,13 @@ def main():
 		db = dbserv.create('sites')
 
 	# start the Tor process
-	start_tor()
-	try:
-		control = stem.control.Controller.from_port(address='127.0.0.1', port=9051)
-	except stem.SocketError as e:
+	handler = TorHandler.TorHandler()
+	if not handler.start_tor():
+		print("[E] There was an error launching Tor. It may already be running.")
+
+	if not handler.start_controller():
 		print("[E] Could not connect to control port!")
-		if DEBUG:
-			print("Error message: {0}".format(e))
 		sys.exit("Please kill all running Tor instances and try again")
-
-	control.authenticate()
-	control.new_circuit(path=None, purpose='general', await_build=False)
-	print(control.get_info('circuit-status'))
-
 		
 	domains = queue.Queue(0)
 	url = check_http(args.url)
@@ -193,10 +119,9 @@ def main():
 
 	# Prints endpoint information if debugging is enabled
 	if DEBUG:
-		print(term.format("[I] Checking Endpoint:", term.Attr.BOLD))
-		print((query("http://www.atagar.com/echo.php").read()).decode('utf8'))
-
-	print(term.format("\nScraping for .ONION domains:\n", term.Attr.BOLD))
+		print(handler.check_endpoint())
+		
+	print("\nScraping for .ONION domains:\n")
 
 	# Sets up DB entry for initial site being scraped
 	if url not in db:	
@@ -211,17 +136,6 @@ def main():
 
 	with concurrent.futures.ThreadPoolExecutor(max_workers=5) as e:
 		while domains.qsize() > 0:
-#			if domains.qsize() > 4:
-#				for x in range(0, 5):
-#					if DEBUG:
-#						print("Thread {0} started!".format(x))
-#					e.submit(scrape_site, domains.get(), domains, db)
-#			else:
-#				length = domains.qsize()
-#				for x in range(0,length):
-#					if DEBUG:
-#						print("Thread {0} started!".format(x))
-#					e.submit(scrape_site, domains.get(), domains, db)
 			scrape_array = []
 			if domains.qsize() > 4:
 				for x in range(0,5):
@@ -230,19 +144,18 @@ def main():
 				for x in range(0,domains.qsize()):
 					scrape_array.append(domains.get())	
 
-			scraper = [e.submit(scrape_site, x, domains, db) for x in scrape_array]
+			scraper = [e.submit(scrape_site, x, domains, db, handler) for x in scrape_array]
 			
 			# Hackish way to make the threads wait until the queue is populated again, or until all threads are done
 			if domains.qsize() == 0:
 				x = [s.result for s in concurrent.futures.as_completed(scraper)]
 				
-		
+	print("\nScraping Complete.")
 
-
-	print(term.format("\nScraping Complete.", term.Attr.BOLD))
-	tor_process.kill()
-	print(term.format("\nTor Instance Killed.", term.Attr.BOLD))
-	
+	if not handler.kill_tor():
+		print("[E] Error killing the Tor process! It may already be dead.")
+	else:
+		print("\nTor Instance Killed.")
 
 if __name__ == "__main__":
 	parser = argparse.ArgumentParser(description="Tor-Scraper.py - Scrape sites using TOR and searches for .onion domains")
